requirements.txt
Markdown==3.5.2
Django==4.2.18
django-filer==3.1.3
django-filter==24.3
django-mptt==0.16.0
django-pagedown==2.2.1
django-tagging==0.5.0
django-taggit==5.0.1
django-taggit-autosuggest==0.4.2
django-taggit-templatetags2==1.6.1
easy-thumbnails==2.10
feedparser==6.0.11
flickrapi==2.4.0
python-dateutil==2.9.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
Pillow==10.4.0
pytz==2024.1
sentry-sdk==2.19.2
requests==2.32.3
twitter==1.19.6
